Get rid of tables layout and rearrange navigation
The only navigation links that work are all and pictures. I've added
words and games to leave room for stories and interactive content in the
future.

I used tables for the layout out of curiousity but that's over now. Each piece
will be delivered inside of an ordered list and each listing will
contain an article with a layout defined by the content specified by the
chosen location.

Images from the pictures model are now presented within figure elements.